fix(calendar): create event on marquee Save-button click
The Save button inside the marquee quick-create popover silently did
nothing — the dialog dismissed, no event was created. Enter worked,
click did not. PR #270 cycled through three patches without success
because all three chased the wrong layer (Popover focus-race, test
flakiness, jsdom virtualization).

Real root cause: the Save button's onClick calls submit(), which
synchronously setIsSubmitting(true). React re-renders with
disabled={isSubmitting} applied to the button *between* the
pointerdown and the click events. The browser then suppresses the
click event on the now-disabled button at the DOM layer — React's
onClick never fires. Enter worked because the input's onKeyDown
handles Enter in the same synchronous event, before any re-render.

Fix:
- Wire submit to onPointerDown (preventDefault on primary button) so
  it runs before the re-render lands; keep onClick as the keyboard-
  activation fallback (Enter/Space on a focused button).
- Replace Radix Popover with Radix Dialog primitive (modal={false},
  positioned at the selection's anchorRect). Explicit-close-only
  semantics, proper a11y (DialogTitle/Description), no auto-dismiss
  on focus changes. Defense in depth even though the primitive was
  not the root cause.
- Swap Popover→Dialog everywhere the marquee popover is rendered:
  day, week, and month views.
- Add regression tests: Save-button click at component, page, and
  E2E levels (day + week view). The pre-existing E2E only exercised
  the Enter path, which is why this bug shipped.
- Update CLAUDE.md gotcha: "submit buttons that disable themselves
  mid-click lose the click".
